# perf: use default memory management within libxml2