@tenderlove 2022-12-20
The performance wins seem nice.  IIRC the reason we used custom allocators was so that the GC had more accurate book keeping WRT to actual memory usage (as well as an opportunity to actually _run_ the GC).  I'd expect this change to cause average memory usage of any particular app to increase since we won't GC as frequently.

That said, it's probably fine to do this.  I just wanted to add some context! 😃