@flavorjones 2022-12-20
@byroot had also asked to re-run the benchmarks with GC disabled, I'll get around to that.