@flavorjones 2022-12-20
With GC disabled ...

HTML4 parsing is 8-25x faster:

    alloc html4 parse 656:    26619.4 i/s
    BASE_ html4 parse 656:    21296.6 i/s - 1.25x  (± 0.07) slower

    alloc html4 parse 70095:      573.4 i/s
    BASE_ html4 parse 70095:      502.9 i/s - 1.14x  (± 0.02) slower

    alloc html4 parse 1929522:       49.1 i/s
    BASE_ html4 parse 1929522:       45.2 i/s - 1.08x  (± 0.02) slower


HTML5 parsing is 0-6x faster:

    alloc html5 parse 656:    25314.1 i/s
    BASE_ html5 parse 656:    23772.3 i/s - 1.06x  (± 0.03) slower

    alloc html5 parse 70095:      280.3 i/s
    BASE_ html5 parse 70095:      273.6 i/s - same-ish: difference falls within error

    alloc html5 parse 1929522:       15.8 i/s
    BASE_ html5 parse 1929522:       15.7 i/s - same-ish: difference falls within error


HTML4 serialization is 10-18x faster:

    alloc html4 srlze 656:    71962.7 i/s
    BASE_ html4 srlze 656:    63734.2 i/s - 1.13x  (± 0.02) slower

    alloc html4 srlze 70095:     1519.9 i/s
    BASE_ html4 srlze 70095:     1286.8 i/s - 1.18x  (± 0.02) slower

    alloc html4 srlze 1929522:      136.5 i/s
    BASE_ html4 srlze 1929522:      123.9 i/s - 1.10x  (± 0.03) slower