@byroot 2022-12-21
Interesting. I would have expected most of the perf difference to be caused by `ruby_xmalloc` triggering GC more frequently, but apparently not. 

👍