@larskanis 2022-12-21
Not triggering the GC can have a quite impressive memory bloat in some situations. In pg.gem we had this issue for years, so that Sam Saffron wrote [a blog post](https://github.com/ged/ruby-pg/commit/c0cad36979a27bf25ef832e51c58dc530ae8b734) about it. To that time I solved this by [adding an estimator](https://github.com/ged/ruby-pg/commit/c0cad36979a27bf25ef832e51c58dc530ae8b734) for the amount of memory that is used behind the scenes and this value is sent to the ruby GC per `rb_gc_adjust_memory_usage()`. This solved the memory bloat issue quite well. Later on a new function to libpq was accepted, that tells the exact amount of memory.

Maybe it's possible to use the native allocation functions in libxml2 and instruct the ruby GC about it per `rb_gc_adjust_memory_usage()`?