@flavorjones 2022-12-21
OK, I'm going to hold off on trying to land this in the upcoming 1.14.0 release, so we can consider a few other solutions.