@casperisfine 2023-03-06
Based on the discovery in https://github.com/sparklemotion/nokogiri/pull/2807, I'm now in support of this change. The reasoning is:

  - It's hard to know for sure which `libxml2` function will allocate or not, so hard to ensure `libxml2` won't allocate during GC.
  - Even if we get it right today, upstream may change and break the gem, causing a vm crash, it's too dangerous.
  - GC trigger heurisitc based on malloc are really not the most relevant ones, it's unlikely to cause increased memory usage in real world scenarios.