@flavorjones 2023-04-07
Closing in favor of #2843 which keeps the current default but allows opting into using system `malloc` via environment variable.